Send the right URL to loaders on fetch page reloads
After a fetch submission we need to reload the routes on the page, which are different routes than those matched by the fetcher.

The code was calling the right `matches` but it was sending the wrong `url`, this led to the routes sending the wrong `?_data` param to the server and then it couldn't send the right data back for the reload.

Now it sends the url of the current page instead of the matched fetcher.